Add extensible "context" argument to the bindings.
Bindings were added various arguments (the whole HTML page, the
content of the HTML tag) and more may be added more in the future
(e.g. the HTML arguments of the tag).  This change ensures that the
code using Weberizer will not need to be modified if these arguments
are added.  Type safety enforces the proper use of these arguments.