rename executable to simply `weberizer`
I'm a stickler for good simple names and don't see why the executable
shouldn't be named simply `weberizer`. Only issue I see is that now
the filename is different from the final executable name, which can be
a bit confusing because you won't know that without reading the _oasis
file. But I still think this is better.