separate lib and app files into separate directories
Actually, main point of this was to let the executable filename be
`weberizer.ml` by putting it in a separate directory from the
Weberizer module implementation, but I'm unable to make that work. I
get a circular build error. I don't see why this shouldn't work. For
now the file is still called `weberizer_compile.ml`.